% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minesweeper.R
\name{minesweeper}
\alias{minesweeper}
\title{Unwind with Minesweeper}
\usage{
minesweeper(gamelog = NULL)
}
\arguments{
\item{gamelog}{An optional specification of existing Minesweeper game log.}
}
\value{
Launches a shiny Gadget to play Minesweeper under user specifications and returns a game log upon closing the shiny Gadget.
}
\description{
Unwind with Minesweeper
}
\examples{
minesweeper()

gamelog_examp <- data.frame(StartTime = Sys.time(),
                            DiffLev = 'Unknown',
                            NumRows = 7,
                            NumCols = 14,
                            NumMines = 10,
                            RandSeed = 123456,
                            GameResult = 'Loss',
                            GameTime = round(difftime(Sys.time()+10, Sys.time(), units = 'secs'), 0))
minesweeper(gamelog_examp)
}
