% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tictactoe.R
\name{tictactoe}
\alias{tictactoe}
\title{Unwind with one-player or two-player Tic-Tac-Toe}
\usage{
tictactoe(gamelog = NULL)
}
\arguments{
\item{gamelog}{An optional specification of existing Tic-Tac-Toe game log.}
}
\value{
Launches a shiny Gadget to play Tic-Tac-Toe under user specifications and returns a game log upon closing the shiny Gadget.
}
\description{
Unwind with one-player or two-player Tic-Tac-Toe
}
\examples{
tictactoe()

gamelog_examp <- data.frame(StartTime = Sys.time(),
                            GameType = 'Unknown',
                            RandSeed = 123456,
                            GameResult = 'Loss',
                            GameTime = round(difftime(Sys.time()+10, Sys.time(), units = 'secs'), 0),
                            NumMoves_X = 0,
                            NumMoves_O = 0)
tictactoe(gamelog_examp)
}
